Refactors to create room for more look up tables
